feat(x): one shared turn renderer for every transcript surface
Extract the chat's conversation-item renderer out of ChatSessionPane into
a shared TurnConversation component and render every turn/chat transcript
through it: full-screen chat, side-pane chat, the sub-agent block's nested
child turn, the bg-task "Runs history" drill-down, and the live-note
"Last run" tab. The bespoke CompactConversation renderer is deleted, so
usage/copy hover behavior, tool rows, and markdown styling are identical
everywhere (the sub-agent transcript's always-visible usage menu came from
it missing the data-turn-usage hover gating).

Quiet tool rows: a settled row now leads with the expand chevron (down
collapsed, up while open — same as the group header) instead of leaving
the glyph slot empty, and the redundant trailing hover chevron is dropped
for those rows. Web search adopts the same lead glyph slot (spinner /
error dot / chevron), aligning its text column with the other rows.

Bg-task detail sidebar switches from the gray bg-sidebar tokens to plain
border/background so run transcripts render on white, matching the
live-note sidebar.
